chore: move state machine enum classes to config package
