+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/configuration/state_machine/negotiation/NegotiationState.java
@@ -1,4 +1,4 @@
-package eu.bbmri.eric.csit.service.negotiator.database.model;
+package eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.negotiation;
 
 public enum NegotiationState {
   SUBMITTED,
