+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/configuration/state_machine/resource/NegotiationIsApprovedGuard.java
@@ -1,6 +1,6 @@
 package eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.resource;
 
-import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationState;
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.negotiation.NegotiationState;
 import eu.bbmri.eric.csit.service.negotiator.service.NegotiationService;
 import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
