+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/database/model/Negotiation.java
@@ -5,6 +5,9 @@ import java.time.ZonedDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 import javax.persistence.*;
+
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.negotiation.NegotiationState;
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.resource.NegotiationResourceState;
 import lombok.*;
 import lombok.ToString.Exclude;
 import org.hibernate.annotations.Type;
