+++ src/test/java/eu/bbmri/eric/csit/service/negotiator/integration/NegotiationLifecycleServiceImplTest.java
@@ -3,11 +3,11 @@ package eu.bbmri.eric.csit.service.negotiator.integration;
 import static org.junit.jupiter.api.Assertions.*;
 
 import eu.bbmri.eric.csit.service.negotiator.NegotiatorApplication;
-import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationEvent;
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.negotiation.NegotiationEvent;
 import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationLifecycleRecord;
-import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationResourceEvent;
-import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationResourceState;
-import eu.bbmri.eric.csit.service.negotiator.database.model.NegotiationState;
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.resource.NegotiationResourceEvent;
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.resource.NegotiationResourceState;
+import eu.bbmri.eric.csit.service.negotiator.configuration.state_machine.negotiation.NegotiationState;
 import eu.bbmri.eric.csit.service.negotiator.database.repository.NegotiationRepository;
 import eu.bbmri.eric.csit.service.negotiator.dto.negotiation.NegotiationCreateDTO;
 import eu.bbmri.eric.csit.service.negotiator.dto.negotiation.NegotiationDTO;
